requirements.txt
uvicorn==0.30.1
transformers==4.42.3
torch==2.3.1
Pillow==10.4.0
requests==2.32.3
matplotlib==3.7.5
chainlit==1.1.305
pydantic-settings==2.3.4
python-dotenv==1.0.1
einops==0.8.0
timm==1.0.7